feat(docsite): shared top nav, separate craft layout
Extract SharedTopNav component used by both docs and craft layouts:
- XDS wordmark in XDSWordmark.tsx
- Nav heading with Craft/Docs menu (XDSNavHeadingMenu)
- Search palette + GitHub button

Route structure:
- (docs)/ route group: DocsShell with side nav + SharedTopNav
- craft/: XDSAppShell with SharedTopNav only (full width, no side nav)

DocsShell simplified — no longer owns top nav, search, or wordmark.
