+++ apps/docsite/src/app/(docs)/packages/[name]/PackageStubPage.tsx
